Added native and library support

+++ instructions/stores/istore.go
@@ -3,41 +3,31 @@ package stores
 import "github.com/go-toy/jvm/instructions/base"
 import "github.com/go-toy/jvm/rtda"
 
-type ISTORE struct {
-	base.Index8Instruction
-}
+type ISTORE struct{ base.Index8Instruction }
 
 func (self *ISTORE) Execute(frame *rtda.Frame) {
-	_istore(frame, self.Index)
+	_istore(frame, uint(self.Index))
 }
 
-type ISTORE_0 struct {
-	base.NoOperandsInstruction
-}
+type ISTORE_0 struct{ base.NoOperandsInstruction }
 
 func (self *ISTORE_0) Execute(frame *rtda.Frame) {
 	_istore(frame, 0)
 }
 
-type ISTORE_1 struct {
-	base.NoOperandsInstruction
-}
+type ISTORE_1 struct{ base.NoOperandsInstruction }
 
 func (self *ISTORE_1) Execute(frame *rtda.Frame) {
 	_istore(frame, 1)
 }
 
-type ISTORE_2 struct {
-	base.NoOperandsInstruction
-}
+type ISTORE_2 struct{ base.NoOperandsInstruction }
 
 func (self *ISTORE_2) Execute(frame *rtda.Frame) {
 	_istore(frame, 2)
 }
 
-type ISTORE_3 struct {
-	base.NoOperandsInstruction
-}
+type ISTORE_3 struct{ base.NoOperandsInstruction }
 
 func (self *ISTORE_3) Execute(frame *rtda.Frame) {
 	_istore(frame, 3)
